services/backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
google-cloud-storage==2.13.0
firebase-admin==6.2.0
channels==4.0.0
channels-redis==4.2.1
gunicorn==21.2.0
python-decouple==3.8
pgvector==0.2.4

drf-spectacular==0.26.5
django-filter==23.3
